REFACTOR CODE

+++ app/src/main/java/com/jeluchu/retrofitapi/WeatherResponse.kt
@@ -14,15 +14,15 @@ class WeatherResponse {
 
 class Main {
     @SerializedName("temp")
-    var temp: Float = 0.toFloat()
+    var temp: Float = 0.0f
     @SerializedName("humidity")
-    var humidity: Float = 0.toFloat()
+    var humidity: Float = 0.0f
     @SerializedName("pressure")
-    var pressure: Float = 0.toFloat()
+    var pressure: Float = 0.0f
     @SerializedName("temp_min")
-    var temp_min: Float = 0.toFloat()
+    var temp_min: Float = 0.0f
     @SerializedName("temp_max")
-    var temp_max: Float = 0.toFloat()
+    var temp_max: Float = 0.0f
 }
 
 class Sys {
